delete old files from 04-high-order-functions directory
