--- 01-syntax/04-high-order-functions/flatten.js
@@ -1,6 +1,0 @@
-var arrays = [[1, 2, 3], [4, 5], [6]];
-
-var result = arrays.reduce(function (i, current) {
-    return i.concat(current);
-}, [])
-console.log(result);
