--- 01-syntax/04-high-order-functions/moduleSupport.js
@@ -1,25 +1,0 @@
-/**
- * Created by Владелец on 14.01.2016.
- */
-function require(name) {
-    if (name in require.cache)
-        return require.cache[name];
-
-    var code = new Function("exports, module", readFile(name));
-    var exports = {}, module = {exports: exports};
-    code(exports, module);
-
-    require.cache[name] = module.exports;
-    return module.exports;
-}
-require.cache = Object.create(null);
-
-
-function readFile(name) {
-    var xhr = new XMLHttpRequest();
-    xhr.open('GET', name + ".js", false);
-    xhr.send(null);
-    return xhr.responseText;
-}
-
-var ancestry = JSON.parse(require('ancestry'));
